Hard-code MAC prefix length (fixes #8)
In a presumed attempt to avoid repeating myself, I had used the length of the
read MAC data from the file, rather than a hard coded length of '10' to decide
how much of the computed MAC to compare with the data.

Since the backup length fields became encrypted, an incorrect passphrase
is likely to lead to an erroneous (and likely very large) length. Since MACs
are embedded after the data payloads, this means that an attempt to read-past
the data will read to the very end of the file. Consequently reading the 10
bytes of expected MAC at this point results in an empty string. Truncating our
computed MAC to the same length yields another empty string and so we
erroneously obtain a match.

Many thanks to @Hildegard49 for finally diagnosing this long-standing bug!